pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
    "pypdf2==3.0.1",
    "pymysql>=1.1.1",
    "cryptography>=45.0.6",
    "pg8000>=1.31.4",
    "python-dotenv>=1.1.1",
    "numpy>=2.3.2",
    "markdown>=3.8.2",
]
